Add support for running the response through a query
Often, the response contains additional items that we don't
care about. In test mode, it can pollute the diff and cause false
positives.
Hence, a query will be useful to remove the unnecessary fields
from the response and then compare them with the contents of the
.ref file.

Question: What happens in the case when someone wants to run
multiple asserts on a response?
For now, they can simply specify their `jq` query in full
to make do. This problem can be solved later.